HBX-1968: Move class DatabaseCollector from 'org.hibernate.tool.api.reveng' to 'org.hibernate.tool.internal.reveng'
