+++ orm/src/main/java/org/hibernate/tool/internal/reveng/AbstractDatabaseCollector.java
@@ -6,7 +6,6 @@ import java.util.Map;
 
 import org.hibernate.mapping.ForeignKey;
 import org.hibernate.tool.api.dialect.MetaDataDialect;
-import org.hibernate.tool.api.reveng.DatabaseCollector;
 import org.hibernate.tool.api.reveng.TableIdentifier;
 
 public abstract class AbstractDatabaseCollector implements DatabaseCollector {
