+++ orm/src/main/java/org/hibernate/tool/internal/reveng/reader/TableCollector.java
@@ -11,10 +11,10 @@ import java.util.Set;
 import org.hibernate.internal.util.StringHelper;
 import org.hibernate.mapping.Table;
 import org.hibernate.tool.api.dialect.MetaDataDialect;
-import org.hibernate.tool.api.reveng.DatabaseCollector;
 import org.hibernate.tool.api.reveng.ReverseEngineeringStrategy;
 import org.hibernate.tool.api.reveng.SchemaSelection;
 import org.hibernate.tool.api.reveng.TableIdentifier;
+import org.hibernate.tool.internal.reveng.DatabaseCollector;
 import org.jboss.logging.Logger;
 
 public class TableCollector {
